requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
python-dotenv==1.0.0
python-docx==1.1.0
openpyxl==3.1.2
